feat: CoreData를 위한 Memo 타입 추가
