+++ CloudNotes/CloudNotes/Model/Memo+CoreDataClass.swift
@@ -1,0 +1,36 @@
+import UIKit.UIFont
+import CoreData
+
+@objc(Memo)
+public class Memo: NSManagedObject {
+  
+}
+
+extension Memo {
+  var subtitle: NSAttributedString {
+    let attributedString = NSMutableAttributedString()
+    
+    if let lastModified = lastModified {
+      let dateFormatter = DateFormatter()
+      dateFormatter.locale = .current
+      dateFormatter.dateStyle = .medium
+      dateFormatter.timeStyle = .none
+      let dateString = dateFormatter.string(from: lastModified)
+      
+      attributedString.append(NSAttributedString(
+        string: dateString + " ",
+        attributes: [.font: UIFont.preferredFont(forTextStyle: .footnote)]
+      ))
+    }
+    let truncatedBody = body?.truncated(limit: 100) ?? ""
+    
+    attributedString.append(NSAttributedString(
+      string: truncatedBody,
+      attributes: [
+        .font: UIFont.preferredFont(forTextStyle: .caption1),
+        .foregroundColor: UIColor.secondaryLabel
+      ]
+    ))
+    return attributedString
+  }
+}
